feat: createOrder - 테스트 실행 전 - Item 도메인에서 import 패키지 관련 빌드 에러 발생

+++ src/main/java/com/sparta/sparta_eats/order/domain/dto/OrderSnapshotDto.java
@@ -2,11 +2,11 @@ package com.sparta.sparta_eats.order.domain.dto;
 
 import lombok.Builder;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 @Builder
-public record OrderSnapshotDto(BigInteger itemTotal,
-                               BigInteger deliveryFee,
-                               BigInteger discountTotal,
-                               BigInteger totalAmount) {
+public record OrderSnapshotDto(BigDecimal itemTotal,
+                               BigDecimal deliveryFee,
+                               BigDecimal discountTotal,
+                               BigDecimal totalAmount) {
 }
